backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-openai
langchain-google-genai
pymupdf
pytesseract
requests
python-dotenv
pytest
httpx
python-docx
openai-whisper
openai-whisper
twilio
